+++ GestiBankBackEnd/src/fr/jgj/gestibank/service/ICompteService.java
@@ -8,7 +8,7 @@ public interface ICompteService {
 	
 	public List<Compte> getAllComptes();
 	
-	public List<Compte> getComptesByClient(long idClient);
+	public List<Compte> getComptesByClient(int idClient);
 	
 	public Compte getCompteByIBAN(String IBAN);
